Move more stuff from stats to Trajectory
