+++ docs/ref/datasets.md
@@ -14,7 +14,6 @@ For more information about the datasets included in evalio, see the [included pi
 ::: evalio.datasets
     options:
         show_root_toc_entry: false
-        show_if_no_docstring: false
         show_labels: false
         filters:
             - "!Dataset"
