+++ docs/ref/types.md
@@ -1,3 +1,16 @@
 ::: evalio.types
     options:
-        show_if_no_docstring: false
+        filters:
+            - "!Trajectory"
+            - "!MetricKind"
+            - "!Metric"
+            - "!Error"
+
+::: evalio.types
+    options:
+        show_root_toc_entry: false
+        members:
+            - Trajectory
+            - MetricKind
+            - Metric
+            - Error
