+++ tests/test_traj_stats.py
